refactor: turn ingestor into module
