+++ graph/Cargo.toml
@@ -3,33 +3,31 @@ name = "trustify-graph"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 trustify-entity = { path = "../entity" }
 trustify-common = { path = "../common" }
 trustify-migration = { path = "../migration" }
 trustify-cvss = { path = "../cvss" }
 
-sea-orm = { version = "0.12", features = ["sea-query-binder", "sqlx-postgres", "runtime-tokio-rustls", "macros", "debug-print"] }
-sea-query = "0.30.0"
-sea-orm-migration = "0.12.2"
-tokio = { version = "1.30.0", features = ["full"] }
+actix-web = "4"
 anyhow = "1.0.72"
-log = "0.4.19"
-env_logger = "0.10.0"
-packageurl = "0.3.0"
-serde_json = "1.0.104"
-spdx-rs = "0.5.3"
-spdx-expression = "0.5.2"
-thiserror = "1"
-csaf = "0.5"
 async-trait = "0.1.74"
-lenient_semver = "0.4.2"
 cpe = "0.1.3"
+csaf = "0.5"
+env_logger = "0.10.0"
+lenient_semver = "0.4.2"
+log = "0.4.19"
+packageurl = "0.3.0"
 postgresql_embedded = { version = "0.6.2", features = ["blocking", "bundled", "tokio"] }
+sea-orm = { version = "0.12", features = ["sea-query-binder", "sqlx-postgres", "runtime-tokio-rustls", "macros", "debug-print"] }
+sea-orm-migration = "0.12.2"
+sea-query = "0.30.0"
+serde_json = "1.0.104"
+spdx-expression = "0.5.2"
+spdx-rs = "0.5.3"
 tempfile = "3"
-
+thiserror = "1"
+tokio = { version = "1.30.0", features = ["full"] }
 
 [dev-dependencies]
 test-log = { version = "0.2.15", features = ["env_logger", "trace"] }
