+++ graph/src/lib.rs
@@ -1,4 +1,5 @@
 #![allow(clippy::module_inception)]
 #![allow(unused)]
 pub mod db;
+pub mod endpoints;
 pub mod graph;
