+++ modules/ingestor/Cargo.toml
@@ -1,15 +1,16 @@
 [package]
-name = "trustify-ingestors"
+name = "trustify-module-ingestor"
 version = "0.1.0"
 edition = "2021"
 publish = false
 
 [dependencies]
-trustify-common = { path = "../common"}
-trustify-graph = { path = "../graph"}
-trustify-entity = { path = "../entity" }
-trustify-cvss = { path = "../cvss" }
+trustify-common = { path = "../../common"}
+trustify-cvss = { path = "../../cvss" }
+trustify-entity = { path = "../../entity" }
+trustify-graph = { path = "../../graph"}
 
+actix-web = "4"
 anyhow = "1.0.72"
 chrono = { version = "0.4.35", features = ["serde"] }
 csaf = "0.5.0"
@@ -25,8 +26,9 @@ serde_json = "1.0.114"
 sha2 = "0.10.8"
 thiserror = "1.0.58"
 tokio = { version = "1.30.0", features = ["full"] }
+utoipa = "4"
+walker-common = "=0.6.0-alpha.8"
 
 [dev-dependencies]
 test-log = { version = "0.2.15", features = ["env_logger", "trace"] }
-
-
+url-escape = "0.1.1"
