+++ modules/ingestor/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod endpoints;
+pub mod service;
