+++ modules/ingestor/src/service/hashing.rs
