+++ server/src/server/read/vulnerability.rs
@@ -1,5 +1,5 @@
-use crate::AppState;
 use actix_web::{get, web, HttpResponse, Responder};
+use trustify_graph::graph::Graph;
 
 #[utoipa::path(
     responses(
@@ -8,7 +8,7 @@ use actix_web::{get, web, HttpResponse, Responder};
 )]
 #[get("vulnerability/{id}/affected/packages")]
 pub async fn affected_packages(
-    state: web::Data<AppState>,
+    state: web::Data<Graph>,
     id: web::Path<String>,
 ) -> actix_web::Result<impl Responder> {
     Ok(HttpResponse::Ok().finish())
@@ -21,7 +21,7 @@ pub async fn affected_packages(
 )]
 #[get("vulnerability/{id}/affected/products")]
 pub async fn affected_products(
-    state: web::Data<AppState>,
+    state: web::Data<Graph>,
     id: web::Path<String>,
 ) -> actix_web::Result<impl Responder> {
     Ok(HttpResponse::Ok().finish())
