--- server/src/server/write/mod.rs
@@ -1,1 +1,0 @@
-pub mod advisory;
